+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
+    "explorer.sortOrder": "mixed",
     "rust-analyzer.check.command": "clippy",
     "[rust]": {
         "editor.defaultFormatter": "rust-lang.rust-analyzer",
